split the project into 2 apps (client, freelancer)
