+++ ProjetPython/settings.py
@@ -35,7 +35,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'myapp.apps.MyappConfig',
+    'client',
+    'freelancer',
     'django_extensions',
 ]
