+++ client/models.py
@@ -1,0 +1,11 @@
+from django.contrib.auth.models import User
+from django.db import models
+
+# Create your models here.
+class Client(models.Model):
+    user = models.OneToOneField(User, on_delete=models.CASCADE)
+    company_name = models.CharField(max_length=255, blank=True)
+    phone_number = models.CharField(max_length=20, blank=True)
+    address = models.TextField(blank=True)
+
+
