--- myapp/templates/myapp/topbar.html
@@ -1,66 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light shadow-sm position-sticky top-0" style="z-index: 1030;">
-    <div class="container-fluid">
-        <a class="navbar-brand fw-bold text-primary" href="{% url 'home' %}">FreelanceHub</a>
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown"
-                aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-
-        {% if not minimal_nav %}
-            <div class="collapse navbar-collapse" id="navbarNavDropdown">
-            <ul class="navbar-nav mx-auto">
-                <li class="nav-item">
-                    <a class="nav-link" href="{% url 'dashboard' %}">Explore</a>
-                </li>
-
-                <!-- Categories dropdown starts here -->
-                <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" id="categoriesDropdown" role="button"
-                       data-bs-toggle="dropdown" aria-expanded="false">
-                        Categories
-                    </a>
-                    <ul class="dropdown-menu" aria-labelledby="categoriesDropdown">
-                        <li><a class="dropdown-item" href="#">Development</a></li>
-                        <li><a class="dropdown-item" href="#">Design</a></li>
-                        <li><a class="dropdown-item" href="#">Writing</a></li>
-                        <li>
-                            <hr class="dropdown-divider">
-                        </li>
-                        <li><a class="dropdown-item" href="#">All Categories</a></li>
-                    </ul>
-                </li>
-                <!-- Categories dropdown ends here -->
-                <li class="nav-item">
-                    <a class="nav-link" href="#how">How it Works</a>
-                </li>
-            </ul>
-
-            <div class="dropdown">
-                <a href="#" class="d-flex align-items-center text-decoration-none dropdown-toggle"
-                   id="profileDropdown" data-bs-toggle="dropdown" aria-expanded="false">
-                    <img src="https://img.freepik.com/free-vector/blue-circle-with-white-user_78370-4707.jpg?t=st=1746195048~exp=1746198648~hmac=638e192329dd951229828bde5bb7891db384b1bcd2496af3ce4f18f6b3ba624f&w=740"
-                         alt="profile"
-                         class="rounded-circle"
-                         style="width: 40px; height: 40px; object-fit: cover;">
-                </a>
-                <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="profileDropdown">
-                    <li><a class="dropdown-item" href="">View Profile</a></li>
-                    <li><a class="dropdown-item" href="{% url 'settings' %}">Settings</a></li>
-                    <li>
-                        <hr class="dropdown-divider">
-                    </li>
-                    <li>
-                        <form method="post" action="">
-                            {% csrf_token %}
-                            <button type="submit" class="dropdown-item text-danger">Logout</button>
-                        </form>
-                    </li>
-                </ul>
-            </div>
-
-        {% endif %}
-        </div>
-        {% endif %}
-    </div>
-</nav>
